I don't know what's changed. I made a Sandbox project. Which probably broke the breakout project. Fixed some issues with the Window class when hosting the window in a WPF application. Line Debug rendering Static rendering class to manually invoke rendering calls.
There's quite a bit to fix and cleanup, but didn't want to continue adding bulk changes. In favor of smaller more precise changes.
Main issue is that I'm not planning anything at the moment. Just working on whatever I feel like, which isn't helping me make any meaningful progress towards an end goal.

+++ Engine/ResourceManager.cpp
@@ -146,7 +146,7 @@ std::shared_ptr<Texture> ResourceManager::LoadTexture(std::string filepath)
 	int32 channels;
 	uint8* data = stbi_load(filepath.c_str(), &width, &height, &channels, 0);
 	std::vector<uint8> pixelData(data, data + (width * height * channels));
-	
+
 	if (data != nullptr)
 	{
 	}
@@ -164,7 +164,7 @@ std::shared_ptr<Texture> ResourceManager::LoadTexture(std::string filepath)
 
 std::shared_ptr<Texture> ResourceManager::CreateColorTexture(glm::u8vec4 color)
 {
-	Guid guid = (color.r << 24) + (color.g << 16) + (color.b << 8) + color.a;
+	Guid guid = ((uint64)color.r << 24) + ((uint64)color.g << 16) + ((uint64)color.b << 8) + color.a;
 	auto existingResource = m_resources.find(guid);
 	if (existingResource != m_resources.end())
 	{
@@ -220,9 +220,9 @@ Guid ResourceManager::GetGuid(std::string name)
 {
 	std::string lowercaseString = name;
 	std::transform(name.begin(), name.end(), lowercaseString.begin(), [](char ch)
-	{
-		return (char)std::tolower((int32)ch);
-	});
+		{
+			return (char)std::tolower((int32)ch);
+		});
 	Guid guid = std::hash<std::string>{}(lowercaseString);
 	return guid;
 }
